create error classes for each api error types
