+++ packages/node-core/src/index.ts
@@ -4,7 +4,7 @@
 import './utils/bigint';
 
 export * from './api.service';
-export * from './api.connection';
+export * from './api.connection.error';
 export * from './logger';
 export * from './profiler';
 export * from './events';
