+++ packages/node-core/src/indexer/connectionPoolState.manager.ts
@@ -4,9 +4,10 @@
 import {OnApplicationShutdown} from '@nestjs/common';
 import chalk from 'chalk';
 import {toNumber} from 'lodash';
-import {IApiConnectionSpecific} from '../api.connection';
+import {ApiErrorType} from '../api.connection.error';
+import {IApiConnectionSpecific} from '../api.service';
 import {getLogger} from '../logger';
-import {errorTypeToScoreAdjustment, ApiErrorType} from './connectionPool.service';
+import {errorTypeToScoreAdjustment} from './connectionPool.service';
 
 const RETRY_DELAY = 60 * 1000;
 const MAX_FAILURES = 5;
